Update class methods for consistency
There was a combination of method names that were confusing. Those
have been normalized so that there get, set, and save for
operating on one object at a time and getMulti, setMulti, and
saveMulti for operating on arrays for objects.

Also, an optional parameter `use_cache` was added to get and
getMulti which causes the Repository to fetch fresh data from
the read handler instead of using values which may be in the
object's storage array. This can be useful if it is known that
some process has modified the data at it's source outside of the
Repository methods.